fix: dynamic load profile page
