+++ src/components/Pages/Profile/index.ts
@@ -1,0 +1,1 @@
+export * from "./Profile";
