+++ src/components/Pages/index.ts
@@ -1,0 +1,2 @@
+export * from "./Homepage";
+export * from "./Profile";
